requirements.txt
firebase_admin
streamlit
streamlit_option_menu
python-dotenv
google-generativeai
scikit-learn 